feat: Support add event and remove event in calendar
