+++ Mori/Services/CalendarMCP.swift
@@ -1,5 +1,6 @@
 import Foundation
 import EventKit
+import SwiftUI
 
 // MARK: - Calendar Response Types
 struct CalendarReadResponse: Codable {
@@ -39,6 +40,14 @@ struct CalendarEvent: Codable {
 struct DateRange: Codable {
     let startDate: String
     let endDate: String
+}
+
+
+
+struct CalendarAddResponse: Codable {
+    let success: Bool
+    let message: String
+    let event: CalendarEvent
 }
 
 struct CalendarUpdateResponse: Codable {
@@ -427,7 +436,7 @@ class CalendarMCP: ObservableObject {
                 isAllDay: isAllDay
             )
             
-            let response = CalendarUpdateResponse(
+            let response = CalendarAddResponse(
                 success: true,
                 message: "Event created successfully",
                 event: calendarEvent
@@ -519,4 +528,442 @@ enum CalendarMCPError: Error, LocalizedError {
             return "Deletion failed: \(message)"
         }
     }
-} 
+}
+
+// MARK: - SwiftUI View Extensions
+extension CalendarMCP {
+    
+    // MARK: - Calendar Read Result View
+    static func createReadResultView(
+        step: WorkflowStep,
+        showingCalendarDetail: Binding<Bool>,
+        showingCalendarConfirmation: Binding<Bool>,
+        eventToOpen: Binding<CalendarEvent?>
+    ) -> some View {
+        Group {
+            if let resultValue = step.details["result"],
+               let jsonData = resultValue.data(using: .utf8),
+               let calendarResponse = try? JSONDecoder().decode(
+                   CalendarReadResponse.self,
+                   from: jsonData
+               )
+            {
+                VStack(spacing: 16) {
+                    // Header with open calendar button
+                    HStack(spacing: 16) {
+                        Image(systemName: "magnifyingglass")
+                            .font(.title2)
+                            .foregroundColor(.white)
+                            .frame(width: 24, height: 24)
+
+                        VStack(alignment: .leading, spacing: 4) {
+                            Text("Found \(calendarResponse.count) events in Calendar")
+                                .font(.body)
+                                .fontWeight(.medium)
+                                .foregroundColor(.white)
+                            
+                            Text(formatDateRange(from: calendarResponse.dateRange))
+                                .font(.caption)
+                                .foregroundColor(.white.opacity(0.7))
+                        }
+
+                        Spacer()
+
+                        Button(action: {
+                            openCalendarApp()
+                        }) {
+                            Image(systemName: "calendar.badge.plus")
+                                .font(.title3)
+                                .foregroundColor(.blue)
+                        }
+                    }
+                    .padding(.horizontal, 20)
+                    .padding(.top, 16)
+
+                    // Events list - show at most 3 events
+                    if !calendarResponse.events.isEmpty {
+                        VStack(spacing: 12) {
+                            // Display first 3 events
+                            ForEach(Array(calendarResponse.events.prefix(3).enumerated()), id: \.offset) { index, event in
+                                CalendarEventDetailRowWithButton(
+                                    event: event,
+                                    onOpenCalendar: {
+                                        eventToOpen.wrappedValue = event
+                                        showingCalendarConfirmation.wrappedValue = true
+                                    }
+                                )
+                            }
+                            
+                            // Show "more" button if there are more than 3 events
+                            if calendarResponse.events.count > 3 {
+                                Button(action: {
+                                    showingCalendarDetail.wrappedValue = true
+                                }) {
+                                    HStack {
+                                        Text("Show \(calendarResponse.events.count - 3) more events")
+                                            .font(.subheadline)
+                                            .fontWeight(.medium)
+                                        
+                                        Image(systemName: "chevron.down")
+                                            .font(.caption)
+                                    }
+                                    .foregroundColor(.blue)
+                                    .padding(.vertical, 12)
+                                    .padding(.horizontal, 16)
+                                    .background(
+                                        RoundedRectangle(cornerRadius: 8)
+                                            .fill(Color.blue.opacity(0.1))
+                                    )
+                                }
+                                .padding(.top, 8)
+                            }
+                        }
+                        .padding(.horizontal, 20)
+                    } else {
+                        VStack {
+                            Image(systemName: "calendar.badge.exclamationmark")
+                                .font(.system(size: 40))
+                                .foregroundColor(.white.opacity(0.5))
+                            Text("No events found")
+                                .font(.subheadline)
+                                .foregroundColor(.white.opacity(0.7))
+                                .padding(.top, 8)
+                        }
+                        .padding(.vertical, 20)
+                    }
+                }
+                .padding(.bottom, 16)
+                .background(
+                    RoundedRectangle(cornerRadius: 16)
+                        .fill(Color.blue.opacity(0.2))
+                )
+                .padding(.horizontal, 20)
+                .alert("Open in Calendar", isPresented: showingCalendarConfirmation) {
+                    Button("Cancel", role: .cancel) { }
+                    Button("Open") {
+                        if let event = eventToOpen.wrappedValue {
+                            openEventInCalendar(event)
+                        }
+                    }
+                } message: {
+                    if let event = eventToOpen.wrappedValue {
+                        Text("Do you want to open '\(event.title)' in the Calendar app?")
+                    }
+                }
+                .sheet(isPresented: showingCalendarDetail) {
+                    CalendarEventsDetailView(
+                        title: "Read Calendar",
+                        subtitle: formatDateRange(from: calendarResponse.dateRange),
+                        events: calendarResponse.events
+                    )
+                }
+            } else {
+                EmptyView()
+            }
+        }
+    }
+    
+    // MARK: - Calendar Update Result View
+    static func createUpdateResultView(
+        step: WorkflowStep,
+        showingCalendarConfirmation: Binding<Bool>,
+        eventToOpen: Binding<CalendarEvent?>
+    ) -> some View {
+        Group {
+            if let resultValue = step.details["result"],
+               let jsonData = resultValue.data(using: .utf8),
+               let updateResponse = try? JSONDecoder().decode(
+                   CalendarUpdateResponse.self,
+                   from: jsonData
+               )
+            {
+                CalendarEventDetailRowWithButton(
+                    event: updateResponse.event,
+                    onOpenCalendar: {
+                        eventToOpen.wrappedValue = updateResponse.event
+                        showingCalendarConfirmation.wrappedValue = true
+                    }
+                )
+                .background(
+                    RoundedRectangle(cornerRadius: 16)
+                        .fill(
+                            (updateResponse.success ? Color.green : Color.red)
+                                .opacity(0.2)
+                        )
+                )
+                .padding(.horizontal, 20)
+                .alert("Open in Calendar", isPresented: showingCalendarConfirmation) {
+                    Button("Cancel", role: .cancel) { }
+                    Button("Open") {
+                        if let event = eventToOpen.wrappedValue {
+                            openEventInCalendar(event)
+                        }
+                    }
+                } message: {
+                    if let event = eventToOpen.wrappedValue {
+                        Text("Do you want to open '\(event.title)' in the Calendar app?")
+                    }
+                }
+            } else {
+                EmptyView()
+            }
+        }
+    }
+    
+    // MARK: - Calendar Add Result View
+    static func createAddResultView(
+        step: WorkflowStep,
+        showingCalendarConfirmation: Binding<Bool>,
+        eventToOpen: Binding<CalendarEvent?>
+    ) -> some View {
+        Group {
+            if let resultValue = step.details["result"],
+               let jsonData = resultValue.data(using: .utf8),
+               let addResponse = try? JSONDecoder().decode(
+                   CalendarAddResponse.self,
+                   from: jsonData
+               )
+            {
+                CalendarEventDetailRowWithButton(
+                    event: addResponse.event,
+                    onOpenCalendar: {
+                        eventToOpen.wrappedValue = addResponse.event
+                        showingCalendarConfirmation.wrappedValue = true
+                    }
+                )
+                .background(
+                    RoundedRectangle(cornerRadius: 16)
+                        .fill(
+                            (updateResponse.success ? Color.green : Color.red)
+                                .opacity(0.2)
+                        )
+                )
+                .padding(.horizontal, 20)
+                .alert("Open in Calendar", isPresented: showingCalendarConfirmation) {
+                    Button("Cancel", role: .cancel) { }
+                    Button("Open") {
+                        if let event = eventToOpen.wrappedValue {
+                            openEventInCalendar(event)
+                        }
+                    }
+                } message: {
+                    if let event = eventToOpen.wrappedValue {
+                        Text("Do you want to open '\(event.title)' in the Calendar app?")
+                    }
+                }
+            } else {
+                EmptyView()
+            }
+        }
+    }
+    
+    // MARK: - Calendar Remove Result View
+    static func createRemoveResultView(
+        step: WorkflowStep
+    ) -> some View {
+        Group {
+            if let resultValue = step.details["result"],
+               let jsonData = resultValue.data(using: .utf8),
+               let deleteResponse = try? JSONDecoder().decode(
+                   CalendarDeleteResponse.self,
+                   from: jsonData
+               )
+            {
+                VStack(spacing: 16) {
+                    // Header with success/failure indicator
+                    HStack(spacing: 16) {
+                        Image(systemName: deleteResponse.success ? "trash.circle" : "xmark.circle")
+                            .font(.title2)
+                            .foregroundColor(.white)
+                            .frame(width: 24, height: 24)
+
+                        VStack(alignment: .leading, spacing: 4) {
+                            Text(deleteResponse.success ? "Event Deleted Successfully" : "Failed to Delete Event")
+                                .font(.body)
+                                .fontWeight(.medium)
+                                .foregroundColor(.white)
+                            
+                            Text(deleteResponse.message)
+                                .font(.caption)
+                                .foregroundColor(.white.opacity(0.7))
+                        }
+
+                        Spacer()
+
+                        if deleteResponse.success {
+                            Button(action: {
+                                openCalendarApp()
+                            }) {
+                                Image(systemName: "calendar")
+                                    .font(.title3)
+                                    .foregroundColor(.blue)
+                            }
+                        }
+                    }
+                    .padding(.horizontal, 20)
+                    .padding(.vertical, 16)
+                }
+                .background(
+                    RoundedRectangle(cornerRadius: 16)
+                        .fill(
+                            (deleteResponse.success ? Color.orange : Color.red)
+                                .opacity(0.2)
+                        )
+                )
+                .padding(.horizontal, 20)
+            } else {
+                EmptyView()
+            }
+        }
+    }
+    
+    // MARK: - Helper Functions
+    private static func openCalendarApp() {
+        let schemes = ["calshow://", "x-apple-calendar://"]
+        
+        for scheme in schemes {
+            print("🔍 Trying to open calendar with scheme: \(scheme)")
+            guard let url = URL(string: scheme) else { 
+                print("❌ Invalid URL scheme: \(scheme)")
+                continue 
+            }
+            
+            if UIApplication.shared.canOpenURL(url) {
+                print("✅ Opening calendar app with: \(scheme)")
+                UIApplication.shared.open(url)
+                return
+            } else {
+                print("❌ Cannot open calendar with scheme: \(scheme)")
+            }
+        }
+        
+        print("❌ No calendar URL schemes worked")
+    }
+    
+    private static func openEventInCalendar(_ event: CalendarEvent) {
+        // Debug logging
+        print("🔍 Trying to open event: \(event.title)")
+        print("🔍 Event ID: '\(event.id)'")
+        print("🔍 Start Date: \(event.startDate)")
+        
+        // Use event ID to open specific event in calendar
+        if !event.id.isEmpty {
+            // Use the correct iOS calendar URL scheme format
+            let eventIdentifier = event.id.replacingOccurrences(of: ":", with: "/")
+            let urlString = "x-apple-calevent://\(eventIdentifier)"
+            
+            print("🔍 Trying URL: \(urlString)")
+            if let url = URL(string: urlString) {
+                if UIApplication.shared.canOpenURL(url) {
+                    print("✅ Opening URL: \(urlString)")
+                    UIApplication.shared.open(url)
+                    return
+                } else {
+                    print("❌ Cannot open URL: \(urlString)")
+                }
+            } else {
+                print("❌ Invalid URL: \(urlString)")
+            }
+        } else {
+            print("❌ Event ID is empty")
+        }
+        
+        // Try to parse the date for calendar URL as fallback
+        let isoFormatter = ISO8601DateFormatter()
+        isoFormatter.timeZone = TimeZone.current
+        
+        if let startDate = isoFormatter.date(from: event.startDate) {
+            // Format date for calendar URL
+            let dateComponents = Calendar.current.dateComponents([.year, .month, .day], from: startDate)
+            if let year = dateComponents.year, let month = dateComponents.month, let day = dateComponents.day {
+                let dateString = String(format: "%04d%02d%02d", year, month, day)
+                let dateUrlString = "calshow://\(dateString)"
+                print("🔍 Trying date URL: \(dateUrlString)")
+                
+                if let dateUrl = URL(string: dateUrlString) {
+                    if UIApplication.shared.canOpenURL(dateUrl) {
+                        print("✅ Opening date URL: \(dateUrlString)")
+                        UIApplication.shared.open(dateUrl)
+                        return
+                    }
+                }
+            }
+        }
+        
+        // Final fallback to opening calendar app
+        print("🔍 Falling back to opening calendar app")
+        openCalendarApp()
+    }
+    
+    // MARK: - Date Range Formatting
+    private static func formatDateRange(from dateRange: DateRange) -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssXXXXX"
+        formatter.timeZone = TimeZone.current
+        
+        // Try to parse start date
+        let startDate: Date?
+        let endDate: Date?
+        
+        let isoFormatter = ISO8601DateFormatter()
+        isoFormatter.timeZone = TimeZone.current
+        
+        // Try ISO format first, then fallback to manual format
+        if let date = isoFormatter.date(from: dateRange.startDate) {
+            startDate = date
+        } else {
+            startDate = formatter.date(from: dateRange.startDate)
+        }
+        
+        if let date = isoFormatter.date(from: dateRange.endDate) {
+            endDate = date
+        } else {
+            endDate = formatter.date(from: dateRange.endDate)
+        }
+        
+        let displayFormatter = DateFormatter()
+        displayFormatter.timeZone = TimeZone.current
+        
+        if let start = startDate, let end = endDate {
+            // Check if it's the same day
+            let calendar = Calendar.current
+            if calendar.isDate(start, inSameDayAs: end) {
+                displayFormatter.dateFormat = "MMM d, yyyy"
+                let dateStr = displayFormatter.string(from: start)
+                
+                // Add time range if not a full day
+                let timeFormatter = DateFormatter()
+                timeFormatter.dateFormat = "HH:mm"
+                timeFormatter.timeZone = TimeZone.current
+                
+                let startTime = timeFormatter.string(from: start)
+                let endTime = timeFormatter.string(from: end)
+                
+                // Check if it's likely a full day (00:00-23:59 or similar)
+                if startTime == "00:00" && (endTime == "23:59" || endTime == "00:00") {
+                    return dateStr
+                } else {
+                    return "\(dateStr) \(startTime)-\(endTime)"
+                }
+            } else {
+                displayFormatter.dateFormat = "MMM d"
+                let startStr = displayFormatter.string(from: start)
+                let endStr = displayFormatter.string(from: end)
+                
+                // Add year if different
+                let yearFormatter = DateFormatter()
+                yearFormatter.dateFormat = "yyyy"
+                let startYear = yearFormatter.string(from: start)
+                let endYear = yearFormatter.string(from: end)
+                
+                if startYear == endYear {
+                    return "\(startStr) - \(endStr), \(startYear)"
+                } else {
+                    return "\(startStr), \(startYear) - \(endStr), \(endYear)"
+                }
+            }
+        }
+        
+        return "Date range: \(dateRange.startDate) - \(dateRange.endDate)"
+    }
+}
