+++ Mori/Services/LLMAIService.swift
@@ -420,6 +420,10 @@ class LLMAIService: ObservableObject {
             return try await calendarMCP.readCalendar(arguments: toolCall.arguments)
         case "update-calendar":
             return try await calendarMCP.updateCalendar(arguments: toolCall.arguments)
+        case "add-calendar":
+            return try await calendarMCP.addCalendar(arguments: toolCall.arguments)
+        case "remove-calendar":
+            return try await calendarMCP.removeCalendar(arguments: toolCall.arguments)
         default:
             throw LLMError.customError("Unknown tool: \(toolCall.tool)")
         }
